add myblog-controller
