+++ app/assets/javascripts/flash.js
@@ -1,0 +1,3 @@
+$(function(){
+  setTimeout("$('.notice, .alert').fadeOut('slow')", 800);
+  });
