+++ app/assets/stylesheets/modules/_blogs.scss
@@ -10,7 +10,7 @@ header {
   line-height: 70px;
   background-color: #194578;
   position: fixed;
-  z-index: 99;
+  z-index: 98;
 }
 
 .header {
