+++ app/helpers/blogs/myblogs_helper.rb
@@ -1,0 +1,2 @@
+module Blogs::MyblogsHelper
+end
