+++ app/views/blogs/index.html.haml
@@ -1,5 +1,4 @@
 .wrapper
- = render "header"
- = render "top-image"
- = render "main"
- = render "footer"
+  = render "top-image"
+  = render "main"
+  = render "footer"
