+++ app/views/blogs/myblogs/index.json.jbuilder
