+++ app/views/blogs/search_show.html.haml
@@ -1,4 +1,3 @@
-= render "header"
 .search
   = render "search"
   .search__right
