+++ app/views/blogs/show.html.haml
@@ -1,4 +1,3 @@
-= render "header"
 .blog
   .blog__left
     .blog__left--container
