+++ app/views/devise/sessions/new.html.haml
@@ -1,4 +1,3 @@
-= render "blogs/header"
 .account-page
   .account-page__inner.clearfix
     .account-page__inner--right.user-form
