+++ app/views/layouts/application.html.haml
@@ -9,4 +9,6 @@
     = stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload'
     = javascript_include_tag 'application', 'data-turbolinks-track': 'reload'
   %body
+    = render "layouts/notifications"
+    = render "blogs/header" 
     = yield
