+++ app/views/users/show.html.haml
@@ -1,4 +1,3 @@
-= render "blogs/header"
 .user
   .user__left
     .user__left__profile
@@ -17,7 +16,7 @@
               %span.user__left__profile__myblog--each--text MyBlog：
               = link_to "#{own_blog.title}", blog_path(own_blog.id), class: "user__profile__myblog--each--content"
               - if user_signed_in? && @user.id == current_user.id
-                = link_to delete_myblog_blog_path(own_blog.id), class: "user__profile__myblog--each--trash", data: {confirm: "MyBlogの登録を解除しますか？"}, method: :delete do
+                = link_to blogs_myblog_path(own_blog.id), class: "user__profile__myblog--each--trash", data: {confirm: "MyBlogの登録を解除しますか？"}, method: :delete do
                   = icon("fa", "trash-alt")
         - else
           .user__left__profile__myblog--caution
@@ -25,7 +24,7 @@
             %span.uset__left__profile__myblog--caution--text MyBlogが未登録です
       - if user_signed_in? && @user.id == current_user.id
         .user__left__profile__myblog--registration
-          = link_to new_myblog_blogs_path, class: "user__left__profile__myblog--registration--icon", data:{"turbolinks" => false} do
+          = link_to new_blogs_myblog_path, class: "user__left__profile__myblog--registration--icon", data:{"turbolinks" => false} do
             = icon("fa", "user-plus")
             %span<>
             %span.user__left__profile__myblog--text MyBlog登録
