+++ config/routes.rb
@@ -1,16 +1,13 @@
 Rails.application.routes.draw do
   devise_for :users
   root "blogs#index"
+  namespace :blogs do
+    resources :myblogs, only: [:index, :new, :create, :destroy]
+  end
   resources :blogs do
     resources :reviews, only: [:create, :edit, :update, :destroy]
     collection do
       get :search_show
-      get :new_myblog
-      get :create_myblog
-      get :search_myblog
-    end
-    member do
-      delete :delete_myblog
     end
   end
