+++ spec/requests/blogs/myblogs_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "Blogs::Myblogs", type: :request do
+
+end
